Bare bones setup with npm scripts
