+++ client/public/manifest.json
@@ -1,0 +1,9 @@
+{
+  "short_name": "console.log",
+  "name": "Social network for web developers",
+  "icons": [],
+  "start_url": ".",
+  "display": "standalone",
+  "theme_color": "#000000",
+  "background_color": "#ffffff"
+}
