+++ client/src/components/AppFooter/AppFooter.js
@@ -1,0 +1,10 @@
+
+const AppFooter = (props) => {
+    return (
+        <div>
+            <sub>Copyright 2022 &cpy;</sub>
+        </div>
+    )
+}
+
+export default AppFooter;
