+++ client/src/components/AppFooter/index.js
@@ -1,0 +1,1 @@
+export { default } from "./AppFooter.js"
