+++ client/src/components/AppHeader/AppHeader.js
@@ -1,0 +1,10 @@
+const AppHeader = (props) => {
+    return (
+        <div>
+            <h1>console.log</h1>
+            <sub>Social network for web developers</sub>
+        </div>
+    );
+}
+
+export default AppHeader;
