+++ client/src/components/AppHeader/index.js
@@ -1,0 +1,1 @@
+export { default } from "./AppHeader.js"
