+++ client/src/pages/Home/Home.js
@@ -1,0 +1,8 @@
+
+const Home = (props) => {
+    return (
+        <h3>Home page</h3>
+    );
+}
+
+export default Home;
